+++ sort.h
@@ -32,6 +32,6 @@ void swap(int *array, int *a, int *b, size_t size);
 int partition(int *array, int start, int end, size_t size);
 void merge(int *array, int *left, int left_size, int *right, int
 right_size);
-void merge_sort(int *array, int size);
+void merge_sort(int *array, size_t size);
 
 #endif
